Fixed add files/libraries selector
VSCE
